feat: connections module
