+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod client;
 pub mod commands;
+pub mod connection;
 pub mod errors;
 pub mod guild;
 pub mod handlers;
